Widen lsb/width to u32 before summing in Sbfx concrete arm
For validated input lsb+width<=64, so the u8 sum never overflows in
practice. But unvalidated callers (raw API users, future fuzz harnesses)
could pass values >127 each and silently wrap through u8, producing
wrong shift_left and a wrong sign-extended result. Match the rest of the
codebase convention (MOVN/MOVZ widen with `as u32` before arithmetic).

Addresses review feedback on PR #143 (claude-review item B3).

+++ src/semantics/concrete.rs
@@ -446,7 +446,10 @@ pub fn apply_instruction_concrete(
             let value = state.get_register(*rn).as_u64();
             // Shift left then arithmetic-right by the same amount, computed on
             // i64, to sign-extend the field MSB across the upper bits.
-            let shift_left = 64 - (*lsb + *width) as u32;
+            // Widen lsb/width to u32 before the sum so we never narrowly wrap
+            // through u8 when the caller passes unvalidated immediates
+            // (validated paths are bounded by is_encodable_aarch64).
+            let shift_left = 64 - ((*lsb as u32) + (*width as u32));
             let intermediate = (value << shift_left) as i64;
             // Right shift by (64 - width) sign-extends from bit (width-1).
             let result = (intermediate >> (64 - *width as u32)) as u64;
